wr-deprecated: add layer
Issue: LINUXEXEC-14994

This layer provides "deprecated" components for Wind River Linux. Because
upstream Yocto project will likely remove support on certain features without
notification, we will have to move these features to this layer to continue
shipping them so that the users will be able to have time to transition to the
replacement functionalities.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>